requirements_web.txt
Flask==2.3.3
spotipy==2.23.0
python-dotenv==1.0.0
Flask-CORS==4.0.0